+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Omit `holdShelfExpirationDate` field in duplicated request. Refs UIREQ-532.
 * Fix sorting by request status. Fixes UIREQ-540.
 * Fix search by tags. Fixes UIREQ-542.
-
+* Fix default pickup service point when editing existing request. Fixes UIREQ-544.
 
 ## [4.0.1](https://github.com/folio-org/ui-requests/tree/v4.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v4.0.0...v4.0.1)